airflow/mwaa-deploy/requirements.txt
dbt-core
dbt-snowflake
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
boto3